Moved helper functions to detail, removed stream from the highest level API
